Switch to using fixtures
