+++ Svc/FpySequencer/CMakeLists.txt
@@ -31,7 +31,7 @@ set(UT_SOURCE_FILES
   "${CMAKE_CURRENT_LIST_DIR}/FpySequencerValidationState.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/test/ut/FpySequencerTester.cpp"
   "${CMAKE_CURRENT_LIST_DIR}/test/ut/FpySequencerTestMain.cpp"
-  "${CMAKE_CURRENT_LIST_DIR}/test/ut/FpySequencerUtils.cpp"
+  "${CMAKE_CURRENT_LIST_DIR}/test/ut/FpySequencerTestSequences.cpp"
 )
 
 register_fprime_ut()
