+++ Svc/FpySequencer/test/ut/FpySequencerTestSequences.cpp
@@ -1,0 +1,24 @@
+#include "FpySequencerTester.hpp"
+
+namespace Svc {
+
+TEST_F(FpySequencerTester, Empty) {
+    allocMem();
+    writeAndRun();
+    dispatchUntilState(State::IDLE);
+    ASSERT_CMD_RESPONSE_SIZE(1);
+    ASSERT_CMD_RESPONSE(0, FpySequencer::OPCODE_RUN, 0, Fw::CmdResponse::OK);
+}
+
+TEST_F(FpySequencerTester, Full) {
+    allocMem();
+    for (U32 ii = 0; ii < Fpy::MAX_SEQUENCE_STATEMENT_COUNT; ii++) {
+        add_NO_OP();
+    }
+    writeAndRun();
+    dispatchUntilState(State::IDLE, Fpy::MAX_SEQUENCE_STATEMENT_COUNT * 4);
+    ASSERT_CMD_RESPONSE_SIZE(1);
+    ASSERT_CMD_RESPONSE(0, FpySequencer::OPCODE_RUN, 0, Fw::CmdResponse::OK);
+}
+
+}
